backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0
sqlalchemy==2.0.20
psycopg2-binary
alembic
openai
numpy